==> noteboooks/01-Emre-Engin-feature-extraction.ipynb <==
# GDATABT
## Ahmet Emre Usta
## Engincan Göçer
### Hacettepe University
### 25.04.2023

import pandas as pd

xls = pd.ExcelFile('Data23.xlsx')
xls.sheet_names

df_sales = pd.read_excel(xls, "SALES")
df_products = pd.read_excel(xls, "PRODUCTS")
df_customers = pd.read_excel(xls, "CUSTOMERS")

df_sales.head()

df_sales.info()

df_customers.head()

df_customers.info()

df_products.head()

df_products.info()

df_products["Category"].unique(), df_products["Category"].value_counts()

df_sales["DateTime"] = pd.to_datetime(df_sales["DateTime"])
df_customers["UserFirstTransaction"] = pd.to_datetime(df_customers["UserFirstTransaction"])


df_sales.describe(datetime_is_numeric=True)

# split the sales data into two dataframes based on the moth of the sales
df_sales_1 = df_sales[df_sales["DateTime"] <= "2017-10-01"].reset_index(drop=True)
df_sales_2 = df_sales[df_sales["DateTime"] > "2017-10-01"].reset_index(drop=True)

df_sales_1.describe(datetime_is_numeric=True)

df_sales_2.describe(datetime_is_numeric=True)

first_6_months = df_sales_1[df_sales_1["DateTime"] <= "2017-07-01"]
last_3_months = df_sales_1[df_sales_1["DateTime"] > "2017-07-01"]


churned_users = []
active_users = []

for user in df_customers["UserID"]:
    if user in first_6_months["UserID"].unique() and user in last_3_months["UserID"].unique():
        active_users.append(user)
        # print(f"User {user} is active")
    if user in first_6_months["UserID"].unique() and user not in last_3_months["UserID"].unique():
        churned_users.append(user)
        # print(f"User {user} is chruned")

# print info about the active and churned users
print("Total users: ", df_customers["UserID"].nunique())
print("9 months users: ", df_sales_1["UserID"].nunique())
print("First 6 months users: ", first_6_months["UserID"].nunique())
print("Last 3 months users: ", last_3_months["UserID"].nunique())
print("Active users: ", len(active_users))
print("Churned users: ", len(churned_users))

#take columns from df_customers which is same as the first_6_months["UserID"] and create a new dataframe
df_train = df_customers[df_customers["UserID"].isin(first_6_months["UserID"])].reset_index(drop=True)
df_train.head()

#add churn column to the df_train dataframe
df_train["isChurn"] = df_train["UserID"].apply(lambda x: 1 if x in churned_users else 0)

#add total price column to the df_train dataframe
df_train["TotalPrice"] = df_train["UserID"].apply(lambda x: first_6_months[first_6_months["UserID"] == x]["Price"].sum())

df_train["TotalSales"] = df_train["UserID"].apply(lambda x: first_6_months[first_6_months["UserID"] == x]["Price"].count())

first_6_months.loc[:, 'Discount'] = first_6_months['Discount'].replace({'Yes': 1, 'No': 0}).astype(int)
df_train["TotalDiscount"] = df_train["UserID"].apply(lambda x: first_6_months[first_6_months["UserID"] == x]["Discount"].sum())

# fid user last transaction date and add it to the df_train dataframe
df_train["UserLastTransaction"] = df_train["UserID"].apply(lambda x: first_6_months[first_6_months["UserID"] == x]["DateTime"].max())

# find user active days from UserLastTransaction - UserFirstTransaction and add it to the df_train dataframe
df_train["UserActiveDays"] = df_train["UserID"].apply(lambda x: (df_train[df_train["UserID"] == x]["UserLastTransaction"] - df_train[df_train["UserID"] == x]["UserFirstTransaction"]).dt.days.values[0])

first_6_months.head()

first_6_months.loc[:, 'Channel'] = first_6_months['Channel'].replace({'WEB': 1, 'MOBILE': 0}).astype(int)
first_6_months["Channel"].unique()

# find user total web channel and add it to the df_train dataframe
df_train["TotalWeb"] = df_train["UserID"].apply(lambda x: first_6_months[(first_6_months["UserID"] == x) & (first_6_months["Channel"] == 1)]["Channel"].count())

df_train["TotalMobile"] = df_train["UserID"].apply(lambda x: first_6_months[(first_6_months["UserID"] == x) & (first_6_months["Channel"] == 0)]["Channel"].count())

first_6_months["PaymentType"].unique()

first_6_months.loc[:, 'PaymentType'] = first_6_months['PaymentType'].replace({'Cash': 1, 'Mobile Payment': 2, 'Online Credit Card': 3}).astype(int)
first_6_months["PaymentType"].unique()

# find user total web channel and add it to the df_train dataframe
df_train["TotalCashPayment"] = df_train["UserID"].apply(lambda x: first_6_months[(first_6_months["UserID"] == x) & (first_6_months["PaymentType"] == 1)]["PaymentType"].count())

df_train["TotalMobilPayment"] = df_train["UserID"].apply(lambda x: first_6_months[(first_6_months["UserID"] == x) & (first_6_months["PaymentType"] == 2)]["PaymentType"].count())

df_train["TotalOnlinePayment"] = df_train["UserID"].apply(lambda x: first_6_months[(first_6_months["UserID"] == x) & (first_6_months["PaymentType"] == 3)]["PaymentType"].count())

# find total days from today to user last transaction and add it to the df_train dataframe
df_train["UserLastTransactionDays"] = df_train["UserID"].apply(lambda x: (pd.to_datetime("today") - df_train[df_train["UserID"] == x]["UserLastTransaction"]).dt.days.values[0])

tmp_df = first_6_months[["UserID","ProductID"]]

merged_df = pd.merge(tmp_df, df_products, on='ProductID', how='left')

category_pivot = pd.pivot_table(merged_df, values='ProductID', index='UserID', columns='Category', aggfunc='count', fill_value=0)


# merge the category_pivot dataframe with the df_train dataframe
df_train = pd.merge(df_train, category_pivot, on='UserID', how='left')
df_train.head()

df_train.to_csv("train.csv", index=False)
