==> noteboooks/02-Emre-xgboost.ipynb <==
# GDATABT
## Ahmet Emre Usta
## Engincan Göçer
### Hacettepe University
### 25.04.2023

import pandas as pd
import numpy as np
import xgboost as xgb
import pickle

from sklearn.preprocessing import LabelEncoder
from sklearn.model_selection import GridSearchCV, StratifiedKFold


df_train = pd.read_csv('train.csv')

df_train.head()

df_train.info()

# print the object type columns
df_train.select_dtypes(include=['object']).columns

# drop UserID column
df_train = df_train.drop(['UserID'], axis=1)

# Train test split
y = df_train['isChurn']
X = df_train.drop(['isChurn'], axis=1)

le = LabelEncoder()
X["Gender"] = le.fit_transform(X["Gender"])

# one hot encoding the Location column
X = pd.get_dummies(X, columns=['Location'])

# convert object type to datetime
X['UserFirstTransaction'] = pd.to_datetime(X['UserFirstTransaction'])
X['UserLastTransaction'] = pd.to_datetime(X['UserLastTransaction'])

# encode the date columns
X["UserFirstTransaction"] = X["UserFirstTransaction"].apply(lambda x: x.toordinal())
X["UserLastTransaction"] = X["UserLastTransaction"].apply(lambda x: x.toordinal())

X.head()

# Define the parameter grid to search over
param_grid = {
    'max_depth': [3, 5, 7],
    'learning_rate': [0.1, 0.01],
    'n_estimators': [50, 100, 200]
}

# Initialize the XGBoost classifier
clf = xgb.XGBClassifier(random_state=42)

# Initialize the GridSearchCV object with the specified parameters
grid_search = GridSearchCV(clf, param_grid, scoring='roc_auc', cv=StratifiedKFold(n_splits=5, shuffle=True, random_state=42))

# Fit the GridSearchCV object to the training data
grid_search.fit(X, y)

# Print the best hyperparameters found by the grid search
print("Best parameters:", grid_search.best_params_)

# Print the best score found by the grid search
print("Best score (ROC AUC):", grid_search.best_score_)


#save the model
pickle.dump(grid_search, open('xgboost.pkl', 'wb'))
