==> noteboooks/04-Emre-test-model.ipynb <==
# GDATABT
## Ahmet Emre Usta
## Engincan Göçer
### Hacettepe University
### 25.04.2023

import pickle
import pandas as pd

from sklearn.preprocessing import LabelEncoder
from sklearn.metrics import accuracy_score, f1_score, roc_auc_score, confusion_matrix

#load the model
model = pickle.load(open('xgboost.pkl', 'rb'))

#run model on test data
df_test = pd.read_csv('test.csv')

df_test.head()

# drop UserID column
df_test = df_test.drop(['UserID'], axis=1)

# Train test split
y_test = df_test['isChurn']
X_test = df_test.drop(['isChurn'], axis=1)

le = LabelEncoder()
X_test["Gender"] = le.fit_transform(X_test["Gender"])

# one hot encoding to categorical features
X_test = pd.get_dummies(X_test, columns=['Location'])

X_test["UserFirstTransaction"] = pd.to_datetime(X_test["UserFirstTransaction"])
X_test["UserLastTransaction"] = pd.to_datetime(X_test["UserLastTransaction"])

# encode the date columns
X_test["UserFirstTransaction"] = X_test["UserFirstTransaction"].apply(lambda x: x.toordinal())
X_test["UserLastTransaction"] = X_test["UserLastTransaction"].apply(lambda x: x.toordinal())

X_test.head()

# Evaluate the best estimator on the test set
y_pred = model.predict(X_test)
print("Test set accuracy:", accuracy_score(y_test, y_pred))
print("Test set F1 score:", f1_score(y_test, y_pred))
print("Test set ROC AUC score:", roc_auc_score(y_test, y_pred))

#print confusion matrix
print(confusion_matrix(y_test, y_pred))
